+++ lib/src/ruleContainer.dart
@@ -1,9 +1,0 @@
-import './validators/baseValidator.dart';
-
-class RuleContainer {
-  List<BaseValidator> rules = new List<BaseValidator>();
-
-  Function getter;
-
-  RuleContainer(this.getter);
-}